requirements.txt
openai-whisper
gradio
requests
